requirements.txt
pyvisa  
numpy  
matplotlib 
telnetlib3
windfreak
requests
pyprl